+++ tests/test_textdirectory.py
@@ -11,6 +11,7 @@ from textdirectory.transformations import transformation_remove_non_ascii, trans
     transformation_to_leetspeak, transformation_crude_spellchecker
 from textdirectory import cli
 
+
 def test_command_line_interface():
     """Test the CLI."""
     runner = CliRunner()
@@ -18,6 +19,7 @@ def test_command_line_interface():
     assert 'TextDirectory' in result.output
     help_result = runner.invoke(cli.main, ['--help'])
     assert 'Usage' in help_result.output
+
 
 def test_simpple_aggregations():
     """Test the simplest form of aggregation."""
@@ -41,6 +43,7 @@ def test_filter_by_similar_documents():
     td.filter_by_similar_documents(reference_file='data/testdata/level_2/Text_E.txt', threshold=0.7)
     assert len(td.aggregation) == 2
 
+
 def test_filter_by_max_filesize():
     """Test the filesize (max) filter."""
     td = TextDirectory(directory='data/testdata/')
@@ -48,12 +51,14 @@ def test_filter_by_max_filesize():
     td.filter_by_max_filesize(max_kb=1)
     assert len(td.aggregation) == 7
 
+
 def test_filter_by_min_filesize():
     """Test the filesize (min) filter."""
     td = TextDirectory(directory='data/testdata/')
     td.load_files(True, 'txt')
     td.filter_by_min_filesize(min_kb=2)
     assert len(td.aggregation) == 1
+
 
 def test_transformation_remove_nl():
     """Test the remove_nl transformation."""
@@ -110,6 +115,16 @@ def test_tabulation(capsys):
     td.print_aggregation()
     out, err = capsys.readouterr()
     assert 'path' in out
+
+
+def test_print_pipeline(capsys):
+    """"Test the print_pipeline function."""
+    td = TextDirectory(directory='data/testdata/')
+    td.load_files(True, 'txt')
+    td.filter_by_chars_outliers()
+    td.print_pipeline()
+    out, err = capsys.readouterr()
+    assert 'filter_by_chars_outliers' in out
 
 
 def test_transform_to_memory():